// src/types/Pet.type.ts
export type Pet = {
  id?: number;
  name: string;
  especie: string;
  porte: string;
  sexo: string;
  faixaEtaria: string;
  nomeTutor: string;
  telefoneTutor: string;
  descricao: string;
  agressivo: boolean;
  brincalhao: boolean;
  carente: boolean;
  castrado: boolean;
  sociavel: boolean;
  vacinado: boolean;
};